Refactor setup_oracle_client
This commit is a near rewrite of the setup_oracle_client script.

+ Running with no arguments downloads a zipfile from Oracle website.
Until recently, these files were not available without logging in so we
had to store local copies.  This is now the easiest default behaviour.
To preserve backwards compatibility, it is still possible to specify
a download location.
+ The libraries are installed into an `instant_client` directory without
a subfolder with a specific version number.  This simplifies writing
scripts and links as the destination of the driver files is known in
advance.  WARNING: this change will break Dockerfile builds where the
LD_LIBRARY_PATH has been hard-coded.
+ The script knows the files required for a complete installation and
will not try to download or install files if they are all present.
+ If an incomplete installation is present, it will be removed before
installation begins.
+ Test coverage is greatly improved.  Fixtures for a dummy zipfile and
a complete installation are used.
+ Logging output in debug mode is improved.